vlm/requirements.txt
fastapi
gdown==4.5.1
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
PyYAML>=5.3.1
opencv-python==4.5.5.64
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
transformers==4.40.2
accelerate
uvicorn[standard]

Pillow==9.1.0
requests